aider: Added a link to the full location page in the location details popup.
# Aider chat conversation:

ASSISTANT: To add a link to the full location page in the location details popup, we need to modify `dashboard/templates/dashboard/pages/map/data.html`.

Here is the *SEARCH/REPLACE* block:

dashboard/templates/dashboard/pages/map/data.html
```html
<<<<<<< SEARCH
            <div class="tags">{{ location.tags }}</div>

            <button class="delete-button" onclick="deletePin(this)"><i class="material-icons">delete</i></button>
=======
            <div class="tags">{{ location.tags }}</div>
            <a href="{% url 'dashboard:view_location' location.id %}" class="view-full-location">View Full Details</a>
            <button class="delete-button" onclick="deletePin(this)"><i class="material-icons">delete</i></button>
>>>>>>> REPLACE
```

This will add a link to the full location page in the location details popup. The link text will be "View Full Details". The URL of the link will be the URL of the location page for the current location.